+++ packages/authentication/modules/access/preAuthenticateJWT/index.ts
@@ -6,14 +6,17 @@ const preAuthenticateJWT = async (token?: string) => {
       throw new Error('No token provided.')
     }
 
-    const response = await fetch(`${process.env.NEXT_PUBLIC_API_BASE_URL}/auth/pre-auth/jwt`, {
-      method: 'POST',
-      body: JSON.stringify({ token }),
-      cache: 'no-store',
-      headers: {
-        'Content-Type': 'application/json',
+    const response = await fetch(
+      `${process.env.NEXT_PUBLIC_API_BASE_URL ?? process.env.EXPO_PUBLIC_API_BASE_URL}/auth/pre-auth/jwt`,
+      {
+        method: 'POST',
+        body: JSON.stringify({ token }),
+        cache: 'no-store',
+        headers: {
+          'Content-Type': 'application/json',
+        },
       },
-    })
+    )
 
     if (response instanceof Response && !response.ok) {
       throw new Error('Failed to pre-authenticate.')