+++ packages/authentication/modules/access/useChangeExpiredPassword/index.ts
@@ -2,11 +2,11 @@ import { setFormApiErrors } from '@baseapp-frontend/utils'
 
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useMutation } from '@tanstack/react-query'
-import { SubmitHandler, useForm } from 'react-hook-form'
+import { type SubmitHandler, useForm } from 'react-hook-form'
 
 import AuthApi from '../../../services/auth'
 import { DEFAULT_INITIAL_VALUES, DEFAULT_VALIDATION_SCHEMA } from './constants'
-import { ChangeExpiredPasswordForm, UseChangeExpiredPassword } from './types'
+import type { ChangeExpiredPasswordForm, UseChangeExpiredPassword } from './types'
 
 const useChangeExpiredPassword = ({
   token,