+++ packages/authentication/modules/access/useLogin/types.ts
@@ -1,11 +1,9 @@
-import { TokenTypes } from '@baseapp-frontend/utils'
-
-import { UseMutationOptions } from '@tanstack/react-query'
-import { UseFormProps } from 'react-hook-form'
+import type { UseMutationOptions } from '@tanstack/react-query'
+import type { UseFormProps } from 'react-hook-form'
 
 import AuthApi from '../../../services/auth'
-import {
-  CustomCookieNames,
+import type {
+  CustomJWTKeyNames,
   LoginMfaRequest,
   LoginRequest,
   LoginResponse,
@@ -13,11 +11,10 @@ import {
 
 type ApiClass = Pick<typeof AuthApi, 'login'>
 
-export interface UseLoginOptions extends CustomCookieNames {
+export interface UseLoginOptions extends CustomJWTKeyNames {
   loginFormOptions?: UseFormProps<Partial<LoginRequest>>
   loginOptions?: UseMutationOptions<LoginResponse, unknown, LoginRequest, any>
   mfaOptions?: UseMutationOptions<LoginResponse, unknown, LoginMfaRequest, any>
-  tokenType?: TokenTypes
   ApiClass?: ApiClass
   enableFormApiErrors?: boolean
 }