+++ packages/authentication/modules/mfa/useMfaActivate/types.ts
@@ -1,7 +1,7 @@
-import { UseMutationOptions } from '@tanstack/react-query'
+import type { UseMutationOptions } from '@tanstack/react-query'
 
 import MfaApi from '../../../services/mfa'
-import { MfaActivationResponse, MfaRequest } from '../../../types/mfa'
+import type { MfaActivationResponse, MfaRequest } from '../../../types/mfa'
 
 type ApiClass = Pick<typeof MfaApi, 'activate'>