+++ packages/core/src/auth/index.tsx
@@ -1,7 +1,0 @@
-export * from './login'
-export * from './signup'
-export * from './recoverPassword'
-export * from './resetPassword'
-export * from './types.d'
-export * from './context'
-export * from './provider'