+++ packages/utils/types/jwt.ts
@@ -1,3 +1,5 @@
+import { ACCESS_KEY_NAME, REFRESH_KEY_NAME } from '../constants/jwt'
+
 export interface JWTResponse {
   access: string
   refresh: string
@@ -9,3 +11,5 @@ export interface JWTContent {
   iat: number
   jti: string
 }
+
+export type JWTKey = typeof ACCESS_KEY_NAME | typeof REFRESH_KEY_NAME