+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @baseapp-frontend/design-system
 
+## 0.0.12
+
+### Patch Changes
+
+- Use `AppRouterCacheProvider` from `@mui/material-nextjs` for emotion cache provider. Remove `NextAppDirEmotionCacheProvider` previous solution.
+- Add common and reusable dependencies to the `pnpm` catalog.
+- Update tsconfig to use `lib.json`.
+
+- Updated dependencies
+  - @baseapp-frontend/utils@3.0.0
+
 ## 0.0.11
 
 ### Patch Changes